cloud_storage: Rename partition_probe.h and probe.h
Rename partition_probe.h(cc) to read_path_probes.h(cc) because we have
two different probes there both containing read-path metrics.

Rename probe.h(cc) to remote_probe.h(cc) because it conataisn only
remote_probe class.

(cherry picked from commit 6c75e9066972c8a5fcb05f664de2931ceb5e8db8)
