+++ src/v/archival/tests/ntp_archiver_test.cc
@@ -16,7 +16,7 @@
 #include "bytes/iostream.h"
 #include "cloud_storage/async_manifest_view.h"
 #include "cloud_storage/fwd.h"
-#include "cloud_storage/partition_probe.h"
+#include "cloud_storage/read_path_probes.h"
 #include "cloud_storage/remote.h"
 #include "cloud_storage/types.h"
 #include "cloud_storage_clients/client_pool.h"
