+++ src/v/cloud_storage/async_manifest_view.cc
@@ -14,7 +14,7 @@
 #include "cloud_storage/logger.h"
 #include "cloud_storage/materialized_resources.h"
 #include "cloud_storage/partition_manifest.h"
-#include "cloud_storage/partition_probe.h"
+#include "cloud_storage/read_path_probes.h"
 #include "cloud_storage/remote.h"
 #include "cloud_storage/spillover_manifest.h"
 #include "cloud_storage/types.h"
