+++ src/v/cloud_storage/read_path_probes.h
