+++ src/v/cloud_storage/remote_probe.h
