+++ src/v/cluster/partition.cc
@@ -12,7 +12,7 @@
 #include "archival/ntp_archiver_service.h"
 #include "archival/upload_housekeeping_service.h"
 #include "cloud_storage/async_manifest_view.h"
-#include "cloud_storage/partition_probe.h"
+#include "cloud_storage/read_path_probes.h"
 #include "cloud_storage/remote_partition.h"
 #include "cluster/logger.h"
 #include "cluster/tm_stm_cache_manager.h"
